Extract bevy stuff into egui widget
