+++ Cargo.toml
@@ -5,6 +5,7 @@ members = [
   "crates/argument_parsers",
   "crates/audio",
   "crates/ball_filter",
+  "crates/bevy_mujoco",
   "crates/bevyhavior_simulator",
   "crates/booster",
   "crates/buffered_watch",
@@ -14,6 +15,7 @@ members = [
   "crates/context_attribute",
   "crates/control",
   "crates/coordinate_systems",
+  "crates/egui_bevy",
   "crates/energy_optimization",
   "crates/filtering",
   "crates/framework",
@@ -29,7 +31,6 @@ members = [
   "crates/kinematics",
   "crates/linear_algebra",
   "crates/motionfile",
-  "crates/bevy_mujoco",
   "crates/nao",
   "crates/object_detection",
   "crates/opn",
@@ -94,6 +95,7 @@ bevy = { version = "0.16.0", default-features = false }
 bevy-inspector-egui = "0.29.1"
 bevy_asset_loader = "0.22.0"
 bevy_egui = "0.33.0"
+bevy_mujoco = { path = "crates/bevy_mujoco" }
 bevy_obj = "0.15"
 bevy_panorbit_camera = { version = "0.26.0", features = ["bevy_egui"] }
 bincode = "1.3.3"
@@ -125,6 +127,7 @@ derive_more = { version = "2.0.1", features = [
 dirs = "6.0.0"
 eframe = { version = "0.31.1", features = ["persistence", "wgpu"] }
 egui = { version = "=0.31.1", features = ["persistence"] }
+egui_bevy = { path = "crates/egui_bevy" }
 egui_dock = { version = "0.16.0", features = ["serde"] }
 egui_extras = { version = "0.31.1", features = ["image"] }
 egui_plot = "0.31.0"
@@ -161,7 +164,6 @@ log = "0.4.25"
 mcap = "0.15.0"
 mlua = { version = "0.10.3", features = ["luajit", "serialize", "vendored"] }
 motionfile = { path = "crates/motionfile" }
-bevy_mujoco = { path = "crates/bevy_mujoco" }
 nalgebra = { version = "0.33.2", features = [
   "convert-glam029",
   "serde",
