+++ crates/egui_bevy/Cargo.toml
@@ -1,0 +1,20 @@
+[package]
+name = "egui_bevy"
+version.workspace = true
+edition.workspace = true
+license.workspace = true
+homepage.workspace = true
+
+[dependencies]
+bevy = { workspace = true, features = [
+  "bevy_render",
+  "bevy_asset",
+  "bevy_core_pipeline",
+  "bevy_gizmos",
+  "bevy_window",
+  "bevy_pbr",
+  "tonemapping_luts",
+] }
+bevy_panorbit_camera = { workspace = true }
+eframe = { workspace = true }
+log = { workspace = true }
